fix(review): consolidate zero-date logic, make exhaustiveness check safe
Four issues from general-reviewer follow-up on the prior commit:

- The new zero-date guard duplicated logic already in
  past-meeting.utils.ts's private parsePastMeetingStartIso. Hoisted both
  into a new, zero-import iso-timestamp.utils.ts (firstValidTimestamp /
  firstValidTimestampMs); getPastMeetingStartTimeMs now delegates to it
  instead of carrying its own copy.

- The past-meeting branch had stopped calling getPastMeetingStartTimeMs
  in favor of the new string-returning helper directly, making the feed
  a fourth, independent read of the same two fields instead of going
  through the canonical helper meetings-dashboard and meeting-organizer
  already use. Restored the call (ms-based, converted to an ISO string
  for the timestamp field); firstValidTimestamp now covers only the
  three sources with no dedicated helper (votes, surveys, documents).

- Restored the deleted comment noting lastMeeting (committee-overview's
  own computed) as the one remaining raw-start_time holdout, since a
  meeting with a zero-date start_time can still make the "Past Meeting"
  card and the feed's top row name different meetings.

- default: assertNever(action) would throw uncaught from a DOM click
  handler on an unrecognized action kind -- in tension with hardening
  the external-url case specifically because LFXV2-1707 will feed
  server-constructed (not just in-app-built) actions into this same
  union. Switched to a compile-time-only exhaustiveness check (a new
  action kind still fails to compile) that no-ops instead of throwing
  at runtime. assertNever itself stays in object.utils.ts for call
  sites where a throw is the right call.

LFXV2-1716
